requirements.txt
absl-py==1.2.0
altair==4.2.0
asttokens==2.1.0
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
camelot-py==0.10.1
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.0
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cmake==3.24.1
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
distro==1.7.0
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.8.0
Flask==2.1.0
flatbuffers==22.10.26
fonttools==4.33.3
ftfy==6.1.1
gast==0.4.0
ghostscript==0.7
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.11.0
idna==3.3
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.11.3
imutils==0.5.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.17.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.3
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mediapipe==0.9.0
nltk==3.7
numpy==1.21.2
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.2
parso==0.8.3
pdfminer.six==20220524
pdftopng==0.2.3
pickleshare==0.7.5
Pillow==9.1.1
pipreqs==0.4.11
plotly==5.11.0
prompt-toolkit==3.0.32
protobuf==3.19.4
pure-eval==0.2.2
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==2.1.0
pyrsistent==0.19.2
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.7.1
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
stack-data==0.6.1
streamlit==1.14.0
tabula==1.0.5
tabula-py==2.4.0
tabulate==0.8.9
tenacity==8.1.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.0.0
tk==0.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.24.0
typing_extensions==4.2.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.1.0
wget==3.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.7.0
